refactor: move sync models to new names and subdir
